Started moving page attachment
