+++ extensions/amp-story-page-attachment/0.1/amp-story-draggable-drawer.js
