+++ extensions/amp-story-page-attachment/OWNERS
@@ -1,0 +1,23 @@
+// For an explanation of the OWNERS rules and syntax, see:
+// https://github.com/ampproject/amp-github-apps/blob/main/owners/OWNERS.example
+
+{
+  rules: [
+    {
+      owners: [
+        {
+          name: 'gmajoulet',
+          notify: true,
+        },
+        {
+          name: 'newmuis',
+          notify: true
+        },
+        {
+          name: 'mszylkowski', 
+          notify: true
+        },
+      ],
+    },
+  ],
+}
