+++ extensions/amp-story/1.0/amp-story-open-page-attachment.js
@@ -2,7 +2,7 @@ import * as Preact from '#core/dom/jsx';
 /**
  * @fileoverview Helper for amp-story rendering of page-attachment UI.
  */
-import {AttachmentTheme} from './amp-story-page-attachment';
+import {AttachmentTheme} from '../../amp-story-page-attachment/0.1/amp-story-page-attachment';
 import {LocalizedStringId_Enum} from '#service/localization/strings';
 import {computedStyle, setImportantStyles} from '#core/dom/style';
 import {dev} from '#utils/log';
