+++ extensions/amp-story/1.0/amp-story.js
@@ -35,7 +35,7 @@ import {AmpStoryEmbeddedComponent} from './amp-story-embedded-component';
 import {AmpStoryGridLayer} from './amp-story-grid-layer';
 import {AmpStoryHint} from './amp-story-hint';
 import {AmpStoryPage, NavigationDirection, PageState} from './amp-story-page';
-import {AmpStoryPageAttachment} from './amp-story-page-attachment';
+import {AmpStoryPageAttachment} from '../../amp-story-page-attachment/0.1/amp-story-page-attachment';
 import {AmpStoryRenderService} from './amp-story-render-service';
 import {AmpStoryShare} from './amp-story-share';
 import {AmpStoryViewerMessagingHandler} from './amp-story-viewer-messaging-handler';
